review

+++ riscv/src/lib.rs
@@ -63,8 +63,8 @@ pub struct CompilerOptions {
     pub field: KnownField,
     pub libs: RuntimeLibs,
     pub continuations: bool,
-    pub min_degree_log: Option<u32>,
-    pub max_degree_log: Option<u32>,
+    pub min_degree_log: Option<u8>,
+    pub max_degree_log: Option<u8>,
 }
 
 impl CompilerOptions {
@@ -98,14 +98,14 @@ impl CompilerOptions {
         }
     }
 
-    pub fn with_min_degree_log(self, log_min_degree: u32) -> Self {
+    pub fn with_min_degree_log(self, log_min_degree: u8) -> Self {
         Self {
             min_degree_log: Some(log_min_degree),
             ..self
         }
     }
 
-    pub fn with_max_degree_log(self, log_max_degree: u32) -> Self {
+    pub fn with_max_degree_log(self, log_max_degree: u8) -> Self {
         Self {
             max_degree_log: Some(log_max_degree),
             ..self
